+++ src/main/kotlin/com/github/andreypfau/intellijton/func/psi/elements.kt
@@ -1,8 +1,7 @@
 package com.github.andreypfau.intellijton.func.psi
 
 import com.github.andreypfau.intellijton.func.FuncIcons
-import com.github.andreypfau.intellijton.func.resolve.FuncFunctionCallIdentifierReference
-import com.github.andreypfau.intellijton.func.resolve.FuncFunctionCallReference
+import com.github.andreypfau.intellijton.func.resolve.*
 import com.github.andreypfau.intellijton.func.stub.FuncFunctionStub
 import com.github.andreypfau.intellijton.func.stub.FuncStubbedNamedElementImpl
 import com.intellij.extapi.psi.ASTWrapperPsiElement
@@ -12,7 +11,6 @@ import com.intellij.openapi.editor.colors.TextAttributesKey
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiNameIdentifierOwner
 import com.intellij.psi.stubs.IStubElementType
-import javax.swing.Icon
 
 interface FuncElement : PsiElement
 abstract class FuncElementImpl(node: ASTNode) : ASTWrapperPsiElement(node), FuncElement
@@ -33,8 +31,8 @@ abstract class FuncFunctionDefinitionMixin : FuncStubbedNamedElementImpl<FuncFun
     constructor(node: ASTNode) : super(node)
     constructor(stub: FuncFunctionStub, nodeType: IStubElementType<*, *>) : super(stub, nodeType)
 
-    override fun getNameIdentifier(): PsiElement? = functionName.identifier
-    override fun getIcon(flags: Int): Icon = FuncIcons.FUNCTION
+    override fun getNameIdentifier() = functionName.identifier
+    override fun getIcon(flags: Int) = FuncIcons.FUNCTION
     override fun getPresentation(): PresentationData {
         val name = name
         return PresentationData(
@@ -54,4 +52,25 @@ abstract class FuncFunctionCallMixin(node: ASTNode) : FuncNamedElementImpl(node)
 abstract class FuncFunctionCallIdentifierMixin(node: ASTNode) : FuncNamedElementImpl(node), FuncFunctionCallIdentifier {
     override fun getNameIdentifier() = identifier
     override fun getReference() = FuncFunctionCallIdentifierReference(this)
+}
+
+abstract class FuncMethodCallMixin(node: ASTNode) : FuncNamedElementImpl(node), FuncMethodCall {
+    override fun getNameIdentifier() = methodCallIdentifier?.identifier
+    override fun getReference() = FuncMethodCallReference(this)
+}
+
+abstract class FuncMethodCallIdentifierMixin(node: ASTNode) : FuncNamedElementImpl(node), FuncMethodCallIdentifier {
+    override fun getNameIdentifier() = identifier
+    override fun getReference() = FuncMethodCallIdentifierReference(this)
+}
+
+abstract class FuncModifyingMethodCallMixin(node: ASTNode) : FuncNamedElementImpl(node), FuncModifyingMethodCall {
+    override fun getNameIdentifier() = modifyingMethodCallIdentifier?.identifier
+    override fun getReference() = FuncModifyingMethodCallReference(this)
+}
+
+abstract class FuncModifyingMethodCallIdentifierMixin(node: ASTNode) : FuncNamedElementImpl(node),
+    FuncModifyingMethodCallIdentifier {
+    override fun getNameIdentifier() = identifier
+    override fun getReference() = FuncModifyingMethodCallIdentifierReference(this)
 }